fix: duplicate includes of Vue
It seems like `@yourname/yourlibrary` has it's own copy of vue in its `node_modules` and as a result Vite is including both in the bundle. The fix here instructs Vite to de-duplicate the library. There may be other approaches to fixing this such has having the library reference a version of vue defined at the workspace level, but the solution in this PR is also fairly conventional.